Clean up the code
Remove reusable components, etc.
