+++ src/screens/Friend/GroupInfoScreen.js
@@ -1,13 +1,5 @@
 import React, { useEffect, useState } from "react";
-import {
-  Image,
-  StyleSheet,
-  ScrollView,
-  Text,
-  TouchableOpacity,
-  View,
-  Alert
-} from "react-native";
+import { Alert } from "react-native";
 
 import { firebase } from "../../firebase/Config";
 import {
@@ -26,7 +18,7 @@ const GroupInfoScreen = ({ navigation, route }) => {
   const [isPendingMemberExpanded, setIsPendingMemberExpanded] = useState(false);
   const groupInfo = route.params.groupData;
 
-  console.log(members);
+  // console.log(members);
 
   useEffect(() => {
     return fetchGroupMembers({
@@ -99,48 +91,3 @@ const GroupInfoScreen = ({ navigation, route }) => {
 };
 
 export default GroupInfoScreen;
-
-const styles = StyleSheet.create({
-  container: {
-    backgroundColor: "darkcyan",
-    padding: 5,
-    flex: 1
-  },
-  contentContainer: {
-    alignItems: "center",
-    justifyContent: "center",
-    marginTop: 10
-  },
-  img: {
-    height: 150,
-    width: 150,
-    borderRadius: 75
-  },
-  name: {
-    fontFamily: Platform.OS === "ios" ? "Gill Sans" : "serif",
-    fontSize: 20,
-    fontWeight: "bold",
-    textAlign: "center",
-    marginVertical: 10
-  },
-  description: {
-    fontSize: 15,
-    fontWeight: "600",
-    color: "aquamarine",
-    textAlign: "center",
-    marginBottom: 10
-  },
-  button: {
-    borderColor: "navy",
-    borderWidth: 1,
-    borderRadius: 10,
-    backgroundColor: "aquamarine",
-    alignSelf: "stretch",
-    margin: 5,
-    padding: 5
-  },
-  buttonText: {
-    textAlign: "center",
-    color: "#2e64e5"
-  }
-});
